# docs/swagger.yaml
basePath: /kitchencontrol/api/v1
definitions:
  shared.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  usecase.FindAllProductOutputDTO:
    properties:
      category:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
...
  usecase.FindProductByIDOutputDTO:
    properties:
      category:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
...
  usecase.KitchenNotifierOutputDTO:
    properties:
      customer_id:
        type: string
      estimated_time:
        type: string
      id:
        type: string
      items:
        items:
          type: string
        type: array
      order_id:
        type: string
      queue:
        description: ordem na fila de preparo
        type: string
      status:
        type: string
    type: object
  usecase.MonitorKitchenOutputDTO:
    properties:
      created_at:
        type: string
      delivery_number:
        type: string
      estimated_time:
        type: string
      id:
        type: string
      order_id:
        type: string
      queue:
        type: string
      status:
        type: string
    type: object
...
  usecase.OrderFindByIdOutputDTO:
    properties:
      created_at:
        type: string
      customercpf:
        type: string
      delivery_number:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/usecase.OrderItemDTO'
        type: array
      status:
        type: string
      total:
        type: number
    type: object
  usecase.OrderFindByParamOutputDTO:
    properties:
      created_at:
        type: string
      customercpf:
        type: string
      delivery_number:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/usecase.OrderItemDTO'
        type: array
      status:
        type: string
      total:
        type: number
    type: object
...
  usecase.OrderItemDTO:
    properties:
      id:
        type: string
      price:
        type: number
      productid:
        type: string
      quantity:
        type: integer
      status:
        type: string
    type: object
  usecase.RegisterProductInputDTO:
    properties:
      category:
        type: string
      description:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
  usecase.RegisterProductOutputDTO:
    properties:
      category:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
  usecase.UpdateProductInputDTO:
    properties:
      category:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
  usecase.UpdateProductOutputDTO:
    properties:
      category:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is fiap fase 2 challenge project.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fiap Fase 2 Challenge Clean Arch API - 9SOAT
  version: "1.0"
paths:
...
  /kitchens/orders/monitor:
    get:
      consumes:
      - application/json
      description: Retorna uma lista de ordens e seus status, ordenado por recebido,
        em preparo e pronto, e por ordem de chegada tbm, e o tempo estimado e o delivery
        number para retirada do pedido
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/usecase.MonitorKitchenOutputDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Busca ordens que estão na cozinha e estão sendo preparadas (ou na fila
        de preparo)
      tags:
      - Kitchens
  /kitchens/orders/notifier:
    post:
      consumes:
      - application/json
      description: Retorna as ordens pagas e não notificadas, notifica a cozinha e
        atualiza o status da ordem para recebida pela cozinha
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/usecase.KitchenNotifierOutputDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Busca ordens pagas e não notificadas, notifica a cozinha e atualiza
        o status da ordem para recebida pela cozinha. A ideia é que algum job chame
        esse endpoint para ficar buscando ordens para a cozinha
      tags:
      - Kitchens
...
  /orders/{id}:
    get:
      consumes:
      - application/json
      description: Get details of a Order and their items by id
      parameters:
      - description: Order id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usecase.OrderFindByIdOutputDTO'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Order not found
          schema:
            type: string
      summary: Get a Order by id
      tags:
      - Orders
...
  /orders/paid:
    get:
      consumes:
      - application/json
      description: Retorna todos os pedidos (orders) registrados no sistema que tenham
        status de pagamento confirmado. Se não houver pedidos, retorna um erro (404).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/usecase.OrderFindByParamOutputDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get all confirmed orders
      tags:
      - Orders
...
  /products:
    get:
      consumes:
      - application/json
      description: Get details of all products
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/usecase.FindAllProductOutputDTO'
            type: array
        "400":
          description: Invalida data
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: No products foundr
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all products
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: Creates a new product in the database
      parameters:
      - description: New Product Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/usecase.RegisterProductInputDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully created
          schema:
            $ref: '#/definitions/usecase.RegisterProductOutputDTO'
        "400":
          description: Invalid data format or missing fields
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "409":
          description: Product already exists
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      summary: Create a new product
      tags:
      - Products
  /products/{id}:
    delete:
      consumes:
      - application/json
      description: Delete details of a Product by id
      parameters:
      - description: Product id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Product deleted
          schema:
            type: string
        "400":
          description: Invalid data
          schema:
            type: string
      summary: Delete a Product
      tags:
      - Products
    get:
      consumes:
      - application/json
      description: Get details of a Product by id
      parameters:
      - description: Product id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usecase.FindProductByIDOutputDTO'
        "404":
          description: Product not found
          schema:
            type: string
      summary: Get a Product
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: Update details of a Product by id
      parameters:
      - description: Product id
        in: path
        name: id
        required: true
        type: string
      - description: Product data
        in: body
        name: Product
        required: true
        schema:
          $ref: '#/definitions/usecase.UpdateProductInputDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Product updated
          schema:
            $ref: '#/definitions/usecase.UpdateProductOutputDTO'
        "400":
          description: Invalid data
          schema:
            type: string
        "404":
          description: Product not found
          schema:
            type: string
      summary: Update a Product
      tags:
      - Products
...
swagger: "2.0"